Split builtin vehicles out of registry into attacks/builtin.py
Separate the catalog mechanism (AttackLibrary + DEFAULT_LIBRARY +
wrap_payload in registry.py) from its contents (the three hand-written
vehicles, now in builtin.py). "builtin" matches the source= provenance
label and contrasts with attacks future adapters import from Garak /
AgentDojo. registry imports BUILTIN_TECHNIQUES; public API unchanged.
